major harmonie runtime and deps reorg
