+++ node/Cargo.toml
@@ -1,16 +1,10 @@
 [package]
-name = "allfeat-node"
-version = "1.0.0-dev"
-description = "The node of the Allfeat chain."
-authors = { workspace = true }
-homepage = "https://allfeat.network"
-edition = "2021"
-license = "Unlicense"
-publish = false
-build = "build.rs"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
+authors.workspace = true
+description = "Allfeat node."
+edition.workspace = true
+name = "alleat"
+readme = "README.md"
+version.workspace = true
 
 [[bin]]
 name = "allfeat"
@@ -61,9 +55,7 @@ sp-timestamp = { workspace = true }
 sp-inherents = { workspace = true }
 sp-offchain = { workspace = true }
 sp-session = { workspace = true }
-frame-system = { workspace = true }
 frame-system-rpc-runtime-api = { workspace = true }
-pallet-transaction-payment = { workspace = true }
 
 allfeat-primitives = { workspace = true }
 
@@ -83,17 +75,17 @@ substrate-frame-rpc-system = { workspace = true }
 pallet-transaction-payment-rpc = { workspace = true }
 pallet-transaction-payment-rpc-runtime-api = { workspace = true }
 pallet-im-online = { workspace = true }
+pallet-staking = { workspace = true }
 
 # Frontier
 fc-api = { workspace = true }
 fc-cli = { workspace = true }
 fc-consensus = { workspace = true }
 fc-db = { workspace = true }
-fc-mapping-sync = { workspace = true }
+fc-mapping-sync = { workspace = true, features = ["sql"] }
 fc-rpc = { workspace = true }
 fc-rpc-core = { workspace = true }
 fc-storage = { workspace = true }
-fp-account = { workspace = true }
 fp-dynamic-fee = { workspace = true }
 fp-rpc = { workspace = true }
 
@@ -103,8 +95,8 @@ frame-benchmarking-cli = { workspace = true }
 
 # Local Dependencies
 shared-runtime = { workspace = true }
-harmonie-runtime = { workspace = true }
-allfeat-runtime = { workspace = true }
+harmonie-runtime = { workspace = true, optional = true }
+allfeat-runtime = { workspace = true, optional = true }
 
 # CLI-specific dependencies
 try-runtime-cli = { workspace = true, optional = true }
@@ -114,28 +106,36 @@ substrate-build-script-utils = { workspace = true }
 frame-benchmarking-cli = { workspace = true, optional = true }
 
 [features]
-default = ["harmonie-native", "allfeat-native", "sql", "txpool", "rocksdb"]
+default = ["harmonie-native", "allfeat-native"]
+
+fast-runtime = [
+    # darwinia
+    "harmonie-runtime?/fast-runtime",
+    "allfeat-runtime?/fast-runtime",
+]
+
 runtime-benchmarks = [
-    "allfeat-runtime/runtime-benchmarks",
-    "harmonie-runtime/runtime-benchmarks",
+    "allfeat-runtime?/runtime-benchmarks",
+    "harmonie-runtime?/runtime-benchmarks",
     "frame-benchmarking-cli/runtime-benchmarks"
 ]
-harmonie-native = []
-allfeat-native = []
+all-natives = [
+    # darwinia
+    "harmonie-native",
+    "allfeat-native",
+]
+harmonie-native = [
+    "harmonie-runtime",
+]
+
+allfeat-native = [
+    "allfeat-runtime",
+]
+
 # Enable features that allow the runtime to be tried and debugged. Name might be subject to change
 # in the near future.
-try-runtime = ["allfeat-runtime/try-runtime", "harmonie-runtime/try-runtime", "try-runtime-cli/try-runtime"]
-rocksdb = [
-    "sc-cli/rocksdb",
-    "sc-service/rocksdb",
-    "fc-cli/rocksdb",
-    "fc-db/rocksdb",
-    "fc-mapping-sync/rocksdb",
-    "fc-rpc/rocksdb",
+try-runtime = [
+    "allfeat-runtime?/try-runtime",
+    "harmonie-runtime?/try-runtime",
+    "try-runtime-cli/try-runtime"
 ]
-sql = [
-    "fc-db/sql",
-    "fc-mapping-sync/sql",
-]
-txpool = ["fc-rpc/txpool"]
-rpc-binary-search-estimate = ["fc-rpc/rpc-binary-search-estimate"]
