+++ node/src/main.rs
@@ -20,7 +20,6 @@
 #![warn(missing_docs)]
 #![allow(clippy::type_complexity, clippy::too_many_arguments, clippy::large_enum_variant)]
 #![cfg_attr(feature = "runtime-benchmarks", warn(unused_crate_dependencies))]
-
 #[cfg(feature = "runtime-benchmarks")]
 mod benchmarking;
 mod chain_specs;
