+++ runtime/harmonie/src/constants.rs
@@ -16,25 +16,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-//! A set of constant values used in substrate runtime.
-/// Money matters.
-pub mod currency {
-	use allfeat_primitives::Balance;
-
-	pub const WEI: Balance = 1;
-	pub const KILOWEI: Balance = 1_000;
-	pub const MEGAWEI: Balance = 1_000_000;
-	pub const GIGAWEI: Balance = 1_000_000_000;
-	pub const MICROAFT: Balance = 1_000_000_000_000;
-	pub const MILLIAFT: Balance = 1_000_000_000_000_000;
-
-	pub const AFT: Balance = 1_000_000_000_000_000_000;
-
-	pub const fn deposit(items: u32, bytes: u32) -> Balance {
-		items as Balance * 20 * AFT + (bytes as Balance) * 100 * MICROAFT
-	}
-}
-
 /// Time.
 pub mod time {
 	use allfeat_primitives::{BlockNumber, Moment};
