+++ runtime/harmonie/src/pallets/scheduler.rs
@@ -16,16 +16,24 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-use allfeat_primitives::Balance;
+use crate::*;
+use frame_support::{parameter_types, traits::EqualPrivilegeOnly};
+use frame_system::EnsureRoot;
 
-pub const WEI: Balance = 1;
-pub const KILOWEI: Balance = 1_000;
-pub const MEGAWEI: Balance = 1_000_000;
-pub const GIGAWEI: Balance = 1_000_000_000;
-pub const MICROAFT: Balance = 1_000_000_000_000;
-pub const MILLIAFT: Balance = 1_000_000_000_000_000;
-pub const AFT: Balance = 1_000_000_000_000_000_000;
+parameter_types! {
+	pub MaximumSchedulerWeight: frame_support::weights::Weight = sp_runtime::Perbill::from_percent(80) *
+		RuntimeBlockWeights::get().max_block;
+}
 
-pub const fn deposit(items: u32, bytes: u32) -> Balance {
-	items as Balance * 10 * AFT + (bytes as Balance) * 100 * MICROAFT
+impl pallet_scheduler::Config for Runtime {
+	type RuntimeEvent = RuntimeEvent;
+	type RuntimeOrigin = RuntimeOrigin;
+	type PalletsOrigin = OriginCaller;
+	type RuntimeCall = RuntimeCall;
+	type MaximumWeight = MaximumSchedulerWeight;
+	type ScheduleOrigin = EnsureRoot<AccountId>;
+	type OriginPrivilegeCmp = EqualPrivilegeOnly;
+	type MaxScheduledPerBlock = ConstU32<50>;
+	type WeightInfo = weights::scheduler::AllfeatWeight<Runtime>;
+	type Preimages = Preimage;
 }
