+++ runtime/shared/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "shared-runtime"
-version = "1.0.0-dev"
-edition = "2021"
-description = "Shared stuff used in runtimes."
 authors.workspace = true
+description = "Shared stuffs for Allfeat runtimes."
+edition.workspace = true
+name = "shared-runtime"
+version.workspace = true
 
 [dependencies]
 enumflags2 = { workspace = true }
@@ -19,7 +19,6 @@ frame-support = { workspace = true }
 
 pallet-identity = { workspace = true }
 pallet-artists = { workspace = true }
-pallet-authorship = { workspace = true }
 pallet-bags-list = { workspace = true }
 pallet-election-provider-multi-phase = { workspace = true }
 pallet-im-online = { workspace = true }
@@ -69,6 +68,8 @@ std = [
     "pallet-timestamp/std",
     "pallet-utility/std",
     "pallet-evm/std",
-    "pallet-authorship/std",
+]
+runtime-benchmarks = [
+    "pallet-identity/runtime-benchmarks"
 ]
 test = []
